+++ python/cugraph/link_analysis/pagerank_wrapper.pyx
@@ -63,7 +63,15 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
 
     cdef bool has_guess = <bool> 0
     if nstart is not None:
-        df['pagerank'][nstart['vertex']] = nstart['values']
+        if len(nstart) != num_verts:
+            raise ValueError('nstart must have initial guess for all vertices')
+        if input_graph.renumbered is True:
+            renumber_series = cudf.Series(input_graph.edgelist.renumber_map.index,
+                                          index=input_graph.edgelist.renumber_map, dtype=np.int32)        
+            vertex_renumbered = renumber_series.loc[nstart['vertex']]
+            df['pagerank'][vertex_renumbered] = nstart['values']
+        else:
+            df['pagerank'][nstart['vertex']] = nstart['values']
         has_guess = <bool> 1
 
     #TODO FIX ME when graph class is upgraded to remove gdf_column
@@ -82,8 +90,10 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
             vertex_renumbered = renumber_series.loc[personalization['vertex']]
             c_pers_vtx = get_column_data_ptr(vertex_renumbered._column)
         else:
-            c_pers_vtx = get_column_data_ptr(personalization['vertex'].astype(np.int32)._column)
-        c_pers_val = get_column_data_ptr(personalization['values'].astype(df['pagerank'].dtype)._column)
+            personalization_vertex = personalization['vertex'].astype(np.int32)
+            c_pers_vtx = get_column_data_ptr(personalization_vertex._column)
+        personalization_values = personalization['values'].astype(df['pagerank'].dtype)
+        c_pers_val = get_column_data_ptr(personalization_values._column)
     
     if (df['pagerank'].dtype == np.float32): 
         c_pagerank.pagerank[int, float](g, <float*> c_pagerank_val, sz, <int*> c_pers_vtx, <float*> c_pers_val,